bug 746975 - update graphite2 library to 1.1.2 release from upstream. r=jdaggett a=akeybl
UltraBlame original commit: e8f573bf8f2e2b4d147a31d37a1ecaa543586a76

+++ gfx/graphite2/src/Code.cpp
@@ -592,7 +592,7 @@ int32 Machine::Code::run(Machine & m, slotref * & map) const
     assert(_own);
     assert(*this);          
 
-    if (m.slotMap().size() <= _max_ref + m.slotMap().context())
+    if (m.slotMap().size() <= size_t(_max_ref + m.slotMap().context()))
     {
         m._status = Machine::slot_offset_out_bounds;
